+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - Runtime commands to enable/disable blurring. Use `chunkc blur::enable` to enable blurring; use `chunkc blur::disable` to disable blurring.
 - Help message. Run `chunkc blur::help` for help about variables and commands.
+- @crisidev created a brew formula!
 
 ### Changed
 - Fix: setting wallpaper on correct screen when working with more than one screen.